refactor: change double to float for MQ135 and DHT11 sensor readings for consistency

+++ src/sensors/sensors.h
@@ -161,8 +161,8 @@ typedef struct s_cfg_t {
   char kvmkey[17]      = "unknown";
   uint8_t log_uart     = 0;
   #ifdef SUPPORT_MQ135
-  double mq135_r0      = 0;
-  double mq135_rl      = 0;
+  float mq135_r0      = 0.0f;
+  float mq135_rl      = 0.0f;
   #endif // SUPPORT_MQ135
   sensor_c_t sensor_cfg[NR_OF_SENSORS] = {0};
 } sensors_cfg_t;
@@ -176,7 +176,7 @@ typedef struct sensor_r_t {
   sensor_c_t *cfg = NULL;
   void   (*init_function)(sensor_r_t*);
   void   (*pre_function)(sensor_r_t*);
-  int8_t (*value_function)(sensor_r_t*, double*);
+  int8_t (*value_function)(sensor_r_t*, float*);
   void   (*post_function)(sensor_r_t*);
   void   (*destroy_function)(sensor_r_t*);
 } sensor_r_t;
@@ -189,8 +189,8 @@ sensors_cfg_t cfg = {
   .kvmkey     = "unknown",
   .log_uart   = 0,
   #ifdef SUPPORT_MQ135
-  .mq135_r0   = 0, // default R0 for MQ-135
-  .mq135_rl   = 0, // default RL for MQ-135
+  .mq135_r0   = 0.0f, // default R0 for MQ-135
+  .mq135_rl   = 0.0f, // default RL for MQ-135
   #endif // SUPPORT_MQ135
   .sensor_cfg = {0}
 };
